Rename `Clipboard::new` to `Clipboard::connect`
