+++ src/platform/windows.rs
@@ -5,7 +5,7 @@ use raw_window_handle::HasRawWindowHandle;
 
 use std::error::Error;
 
-pub fn new_clipboard<W: HasRawWindowHandle>(
+pub fn connect<W: HasRawWindowHandle>(
     _window: &W,
 ) -> Result<Box<dyn ClipboardProvider>, Box<dyn Error>> {
     Ok(Box::new(Clipboard))
